requirements.txt
psycopg2-binary
sqlalchemy
sqlalchemy-redshift
sqlalchemy_mate
attrs
attrs_mate==0.0.4
pyyaml
pathlib_mate==0.0.15
rolex==0.0.8
sfm==0.0.16
numpy
pandas
